requirements.txt
-i https://pypi.org/simple
asgiref==3.9.1; python_version >= '3.9'
attrs==25.3.0; python_version >= '3.8'
django==5.2.5; python_version >= '3.10'
django-tenants==3.8.0
djangorestframework==3.16.1; python_version >= '3.9'
drf-spectacular==0.28.0; python_version >= '3.7'
gunicorn==23.0.0; python_version >= '3.7'
inflection==0.5.1; python_version >= '3.5'
iniconfig==2.1.0; python_version >= '3.8'
jsonschema==4.25.0; python_version >= '3.9'
jsonschema-specifications==2025.4.1; python_version >= '3.9'
packaging==25.0; python_version >= '3.8'
pluggy==1.6.0; python_version >= '3.9'
psycopg2-binary==2.9.10; python_version >= '3.8'
pygments==2.19.2; python_version >= '3.8'
pytest==8.4.1; python_version >= '3.9'
pytest-django==4.11.1; python_version >= '3.8'
python-dotenv==1.1.1; python_version >= '3.9'
pyyaml==6.0.2; python_version >= '3.8'
referencing==0.36.2; python_version >= '3.9'
rpds-py==0.27.0; python_version >= '3.9'
sqlparse==0.5.3; python_version >= '3.8'
uritemplate==4.2.0; python_version >= '3.9'
